added changes to App and renamed folder
