+++ client/src/App.js
@@ -12,7 +12,7 @@ import Home from './pages/Home';
 import Signup from './pages/Signup';
 import Login from './pages/Login';
 import SingleThought from './pages/SingleThought';
-import Header from './components/Header';
+import Header from './components/Navbar';
 import Footer from './components/Footer';
 
 // Construct our main GraphQL API endpoint
