+++ client/src/components/Navbar/navbar.scss
@@ -1,0 +1,31 @@
+.navbar{
+    position: sticky;
+    display: flex;
+    flex-direction: row;
+    height: 8vh
+}
+
+// .logo,
+// .page-links,
+// .user-links{
+//     border: 2px dashed black;
+// }
+
+.logo,
+.user-links,
+.page-links{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.page-links{
+    flex-grow: 10;
+}
+
+.user-link,
+.page-link{
+    margin: 0 1rem;
+    text-decoration: none;
+    color: black;
+}
